move From<Instruction> for ProgramRom into program/columns.rs
