+++ circuits/src/program/columns.rs
@@ -1,4 +1,9 @@
+use itertools::izip;
+use plonky2::hash::hash_types::RichField;
+
 use crate::columns_view::{columns_view_impl, make_col_map, NumberOfColumns};
+use crate::cpu::columns::Instruction;
+use crate::generation::ascending_sum;
 use crate::linear_combination::Column;
 use crate::linear_combination_typed::ColumnWithTypedInput;
 use crate::stark::mozak_stark::{ProgramTable, TableWithTypedOutput};
@@ -18,6 +23,35 @@ pub struct ProgramRom<T> {
     /// - `rs1_select`, `rs2_select`, and `rd_select`
     /// - `imm_value`
     pub inst_data: T,
+}
+
+impl<F: RichField> From<Instruction<F>> for ProgramRom<F> {
+    fn from(inst: Instruction<F>) -> Self {
+        pub fn reduce_with_powers<F: RichField, I: IntoIterator<Item = F>>(
+            terms: I,
+            alpha: u64,
+        ) -> F {
+            izip!((0..).map(|i| F::from_canonical_u64(alpha.pow(i))), terms)
+                .map(|(base, val)| base * val)
+                .sum()
+        }
+
+        Self {
+            pc: inst.pc,
+            inst_data: reduce_with_powers(
+                [
+                    ascending_sum(inst.ops),
+                    inst.is_op1_signed,
+                    inst.is_op2_signed,
+                    inst.rs1_selected,
+                    inst.rs2_selected,
+                    inst.rd_selected,
+                    inst.imm_value,
+                ],
+                1 << 5,
+            ),
+        }
+    }
 }
 
 // Total number of columns.
